fix cast error

+++ android/src/main/java/com/oney/WebRTCModule/GetUserMediaImpl.java
@@ -322,8 +322,8 @@ class GetUserMediaImpl {
 
     private VideoTrack createScreenTrack() {
         DisplayMetrics displayMetrics = getDisplayMetrics();
-        int width = displayMetrics.widthPixels * 0.5;
-        int height = displayMetrics.heightPixels * 0.5;
+        int width = (int) (displayMetrics.widthPixels * 0.5);
+        int height = (int) (displayMetrics.heightPixels * 0.5);
         int fps = 30;
         ScreenCaptureController screenCaptureController = new ScreenCaptureController(width, height, fps, mediaProjectionPermissionResultData);
         return createVideoTrack(screenCaptureController);
